test_iris_systems.py
#!/usr/bin/env python3
"""
Test script for IRIS Robot systems
Tests the satellite image analysis and navigation systems
"""

import os
import sys
import tempfile
import numpy as np
from PIL import Image
import logging

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def test_satellite_analysis():
    """Test the satellite image analysis system"""
    logger.info("Testing Satellite Image Analysis System...")
    
    try:
        # Import the satellite analyzer
        from codesnip3 import IRISSatelliteAnalyzer
        
        # Create a test image
        test_image = create_test_satellite_image()
        
        # Initialize analyzer
        analyzer = IRISSatelliteAnalyzer()
        
        # Test image preprocessing
        processed_image = analyzer.preprocess_image(test_image)
        assert processed_image.shape == (1, 224, 224, 3), f"Unexpected image shape: {processed_image.shape}"
        logger.info("✓ Image preprocessing test passed")
        
        # Test model creation
        assert analyzer.model is not None, "Model was not created"
        logger.info("✓ Model creation test passed")
        
        # Test image analysis (this will use the untrained model)
        try:
            results = analyzer.analyze_image(test_image)
            assert 'classification' in results, "Analysis results missing classification"
            assert 'water_harvesting_potential' in results, "Analysis results missing water potential"
            logger.info("✓ Image analysis test passed")
        except Exception as e:
            logger.warning(f"Image analysis test failed (expected for untrained model): {e}")
        
        # Test batch analysis
        test_dir = create_test_image_directory()
        try:
            batch_results = analyzer.batch_analyze(test_dir)
            assert isinstance(batch_results, list), "Batch analysis should return a list"
            logger.info("✓ Batch analysis test passed")
        except Exception as e:
            logger.warning(f"Batch analysis test failed (expected for untrained model): {e}")
        
        # Cleanup
        cleanup_test_files(test_dir)
        
        logger.info("✓ Satellite Analysis System tests completed successfully")
        return True
        
    except Exception as e:
        logger.error(f"✗ Satellite Analysis System test failed: {e}")
        return False

def test_navigation_system():
    """Test the navigation system"""
    logger.info("Testing Navigation System...")
    
    try:
        # Import the navigation system
        from codesnip4 import IRISNavigationSystem, Point, Obstacle
        
        # Test Point class
        p1 = Point(0, 0)
        p2 = Point(3, 4)
        distance = p1.distance_to(p2)
        assert abs(distance - 5.0) < 0.001, f"Distance calculation error: {distance}"
        logger.info("✓ Point class test passed")
        
        # Test Obstacle class
        obstacle = Obstacle(
            position=Point(1, 1),
            radius=0.5,
            confidence=0.8,
            obstacle_type='static',
            timestamp=0.0
        )
        assert obstacle.position.x == 1.0, "Obstacle position error"
        logger.info("✓ Obstacle class test passed")
        
        # Test Navigation System initialization
        nav_system = IRISNavigationSystem(use_simulation=True)
        assert nav_system.lidar is not None, "LIDAR simulator not initialized"
        assert nav_system.obstacle_detector is not None, "Obstacle detector not initialized"
        assert nav_system.navigation_planner is not None, "Navigation planner not initialized"
        logger.info("✓ Navigation System initialization test passed")
        
        # Test LIDAR simulation
        distances = nav_system.lidar.get_scan_data()
        assert len(distances) > 0, "LIDAR simulation returned empty data"
        logger.info("✓ LIDAR simulation test passed")
        
        # Test obstacle detection
        obstacles = nav_system.obstacle_detector.detect_obstacles(
            nav_system.lidar.get_scan_data(),
            nav_system.lidar.angles
        )
        assert isinstance(obstacles, list), "Obstacle detection should return a list"
        logger.info("✓ Obstacle detection test passed")
        
        # Test navigation planning
        nav_system.navigation_planner.set_goal(5.0, 3.0)
        command = nav_system.navigation_planner.plan_path(obstacles)
        assert hasattr(command, 'linear_velocity'), "Navigation command missing linear velocity"
        assert hasattr(command, 'angular_velocity'), "Navigation command missing angular velocity"
        logger.info("✓ Navigation planning test passed")
        
        logger.info("✓ Navigation System tests completed successfully")
        return True
        
    except Exception as e:
        logger.error(f"✗ Navigation System test failed: {e}")
        return False

def create_test_satellite_image():
    """Create a test satellite image for testing"""
    # Create a simple test image (224x224 RGB)
    image_array = np.random.randint(0, 255, (224, 224, 3), dtype=np.uint8)
    
    # Add some structure to make it look more like a satellite image
    # Create a simple pattern
    for i in range(224):
        for j in range(224):
            # Add some horizontal lines (like terrain features)
            if i % 50 < 10:
                image_array[i, j] = [100, 150, 100]  # Greenish
            # Add some vertical lines
            elif j % 50 < 10:
                image_array[i, j] = [150, 100, 100]  # Reddish
    
    return image_array

def create_test_image_directory():
    """Create a temporary directory with test images"""
    temp_dir = tempfile.mkdtemp(prefix="iris_test_")
    
    # Create a few test images
    for i in range(3):
        image_array = create_test_satellite_image()
        image = Image.fromarray(image_array)
        image_path = os.path.join(temp_dir, f"test_image_{i}.jpg")
        image.save(image_path)
    
    return temp_dir

def cleanup_test_files(test_dir):
    """Clean up test files"""
    try:
        import shutil
        shutil.rmtree(test_dir)
    except Exception as e:
        logger.warning(f"Could not cleanup test directory: {e}")

def test_dependencies():
    """Test that all required dependencies are available"""
    logger.info("Testing Dependencies...")
    
    required_packages = [
        'tensorflow',
        'cv2',  # opencv-python
        'numpy',
        'matplotlib',
        'PIL',  # Pillow
        'sklearn',
        'pandas'
    ]
    
    missing_packages = []
    
    for package in required_packages:
        try:
            if package == 'cv2':
                import cv2
            elif package == 'PIL':
                from PIL import Image
            else:
                __import__(package)
            logger.info(f"✓ {package} is available")
        except ImportError:
            logger.warning(f"✗ {package} is missing")
            missing_packages.append(package)
    
    if missing_packages:
        logger.error(f"Missing packages: {missing_packages}")
        logger.error("Install missing packages with: pip install -r requirements.txt")
        return False
    
    logger.info("✓ All dependencies are available")
    return True

def main():
    """Run all tests"""
    logger.info("Starting IRIS Robot System Tests...")
    logger.info("=" * 50)
    
    # Test dependencies first
    if not test_dependencies():
        logger.error("Dependency test failed. Please install missing packages.")
        return 1
    
    # Test satellite analysis system
    satellite_test_passed = test_satellite_analysis()
    
    # Test navigation system
    navigation_test_passed = test_navigation_system()
    
    # Summary
    logger.info("=" * 50)
    logger.info("TEST SUMMARY:")
    logger.info(f"Satellite Analysis System: {'✓ PASSED' if satellite_test_passed else '✗ FAILED'}")
    logger.info(f"Navigation System: {'✓ PASSED' if navigation_test_passed else '✗ FAILED'}")
    
    if satellite_test_passed and navigation_test_passed:
        logger.info("🎉 All tests passed! IRIS Robot systems are ready to use.")
        return 0
    else:
        logger.error("❌ Some tests failed. Please check the error messages above.")
        return 1

if __name__ == "__main__":
    exit(main()) 