requirements.txt
# IRIS Robot - Python Dependencies
# Core dependencies for satellite image analysis and navigation

# Machine Learning and Computer Vision
tensorflow>=2.10.0
opencv-python>=4.6.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0

# Data processing and utilities
scikit-learn>=1.1.0
pandas>=1.4.0

# Optional: GPU acceleration (uncomment if you have CUDA)
# tensorflow-gpu>=2.10.0

# Optional: Additional ML libraries
# torch>=1.12.0
# torchvision>=0.13.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Logging and monitoring
# Already included in Python standard library

# Note: For Arduino code (codesnip1.cpp and codesnip2.cpp), 
# install Arduino IDE and the following libraries:
# - DHT sensor library
# - LiquidCrystal I2C
# - Wire (built-in)
# - EEPROM (built-in) 